+++ qiskit/providers/aer/backends/unitary_simulator.py
@@ -87,9 +87,12 @@ class UnitarySimulator(AerBackend):
         'conditional': False,
         'open_pulse': False,
         'memory': False,
-        'max_shots': 1,
-        'description': 'A Python simulator for computing the unitary'
-                       'matrix for experiments in qobj files',
+        'max_shots': int(1e6),  # Note that this backend will only ever
+                                # perform a single shot. This value is just
+                                # so that the default shot value for execute
+                                # will not raise an error when trying to run
+                                # a simulation
+        'description': 'A C++ unitary simulator for QASM Qobj files',
         'coupling_map': None,
         'basis_gates': [
             'u1', 'u2', 'u3', 'cx', 'cz', 'id', 'x', 'y', 'z', 'h', 's', 'sdg',