+++ CHANGELOG.md
@@ -57,9 +57,19 @@ Added
 
 Changed
 -------
+- Moved the location of several functions (\#568):
+  - Moved contents of `qiskit.provider.aer.noise.errors` into the `qiskit.providers.noise` module
+  - Moved contents of `qiskit.provider.aer.noise.device` into the `qiskit.providers.noise` module.
+  - Moved contents of `qiskit.provider.aer.noise.utils` into the `qiskit.provider.aer.utils` module.
+
+Deprecated
+----------
+- Deprecated `utils.qobj_utils` functions (\#568)
 
 Removed
 -------
+- Removed `NoiseModel.as_dict`, `QuantumError.as_dict`, `ReadoutError.as_dict`, and 
+  `QuantumError.kron` methods that were deprecated in 0.3 (\#568).
 
 Fixed
 -----